+++ src/melee/ty/toy.c
@@ -65,7 +65,7 @@ typedef struct ToyUnkJObjData {
 /* 3062EC */ static void _Toy_803062EC(s32 arg0, u32 arg1, f32 farg0);
 /* 3064B8 */ static s16 _Toy_803064B8(s16 arg0, s8 arg1);
 /* 30663C */ static void _Toy_8030663C(void);
-/* 306A0C */ static void _Toy_80306A0C(HSD_GObj* arg0, int);
+/* 306A0C */ static void _Toy_80306A0C(HSD_GObj* arg0, intptr_t);
 /* 306C5C */ static void _Toy_80306C5C(HSD_GObj* arg0);
 /* 307018 */ static void _Toy_80307018(void);
 /* 30715C */ static void _Toy_8030715C(f32, f32);
@@ -89,7 +89,7 @@ typedef struct ToyUnkJObjData {
 /* 3114E8 */ static void _Toy_803114E8(void);
 /* 311788 */ static void _Toy_80311788(void);
 /* 311F5C */ static void _Toy_80311F5C(void);
-/* 312050 */ static void _Toy_80312050(HSD_GObj* gobj, int code);
+/* 312050 */ static void _Toy_80312050(HSD_GObj* gobj, intptr_t code);
 
 /// All trophies unlocked?
 bool un_80304470(void)
@@ -1759,7 +1759,7 @@ void Toy_803067BC(s32 arg0, s32 arg1)
     }
 }
 
-void Toy_803068E0(HSD_GObj* gobj, int unused)
+void Toy_803068E0(HSD_GObj* gobj, intptr_t unused)
 {
     if (HSD_CObjSetCurrent(gobj->hsd_obj) != 0) {
         HSD_GObj_80390ED0(gobj, 7);
@@ -1768,12 +1768,12 @@ void Toy_803068E0(HSD_GObj* gobj, int unused)
     }
 }
 
-void Toy_80306930(HSD_GObj* gobj, int unused)
+void Toy_80306930(HSD_GObj* gobj, intptr_t unused)
 {
     HSD_FogSet(gobj->hsd_obj);
 }
 
-void Toy_80306954(HSD_GObj* gobj, int unused)
+void Toy_80306954(HSD_GObj* gobj, intptr_t unused)
 {
     void* state;
     s32 light_idx;
@@ -1796,7 +1796,7 @@ void Toy_80306954(HSD_GObj* gobj, int unused)
     }
 }
 
-void _Toy_80306A0C(HSD_GObj* arg0, int unused)
+void _Toy_80306A0C(HSD_GObj* arg0, intptr_t unused)
 {
     HSD_Fog* fog = GET_FOG(arg0);
 
@@ -6259,7 +6259,7 @@ void Toy_Scene_OnFrame(void)
     }
 }
 
-void _Toy_80312050(HSD_GObj* gobj, int code)
+void _Toy_80312050(HSD_GObj* gobj, intptr_t code)
 {
     HSD_CObj* cobj;
     TyViewData* data;
